backend/requirements.txt
fastapi
uvicorn
gunicorn
azure-identity
azure-keyvault-secrets
azure-monitor-opentelemetry